UPDATED PACKAGES

+++ package.js
@@ -1,22 +1,56 @@
-#!/bin/env node
+#!/usr/bin/env node
 import shell from 'shelljs'
 import exec from 'shelljs.exec'
 import archiver from 'archiver'
 import fs from 'fs'
+import process from 'process'
 
-const dist  = './dist'
-shell.rm('-rf',dist)
-shell.mkdir('-p',dist)
-exec(`yarn build`)
+(async() => {
+   const args = process.argv.slice(2)
+   if (args.length === 0 ) {
+      console.log('no arguments, packaging for deployment')
+      const dist  = './dist'
+      shell.rm('-rf',dist)
+      shell.mkdir('-p',dist)
+      exec(`yarn build`)
 
-const cdir=shell.pwd()
-exec(`open -n -b com.google.chrome --args --pack-extension=${cdir}/dist --pack-extension-key=${cdir}/dist.pem`)
+      const cdir=shell.pwd()
+      exec(`open -n -b com.google.chrome --args --pack-extension=${cdir}/dist --pack-extension-key=${cdir}/dist.pem`)
 
-const DEFZIP  = "gridlinks.zip"
-const output = fs.createWriteStream(DEFZIP)
-const archive = archiver('zip')
-archive.pipe(output)
-archive.directory('dist', false)
-archive.file('auto-update.xml', false)
-archive.file('dist.crx', false)
-await archive.finalize()
+      const DEFZIP  = "gridlinks.zip"
+      const output = fs.createWriteStream(DEFZIP)
+      const archive = archiver('zip')
+      archive.pipe(output)
+      archive.directory('dist', false)
+      archive.file('auto-update.xml', false)
+      archive.file('dist.crx', false)
+      await archive.finalize()
+      console.log(`archive created: ${DEFZIP}`)
+   } else {
+      if(args[0] === 'patch') {
+         const nargs='npx ncu -u -t patch'
+         console.log(nargs)
+         console.log(await exec(nargs).stdout)
+
+      } else if(args[0] === 'minor') {
+         const nargs='npx ncu -u -t minor'
+         console.log(nargs)
+         console.log(await exec(nargs).stdout)
+
+      } else if(args[0] === '-v') {
+         const nargs='npx ncu'
+         console.log(nargs)
+         console.log(await exec(nargs).stdout)
+
+      } else if(args[0] === '-f') {
+         const nargs='npx ncu -u -f '+args[1]
+         console.log(nargs)
+         console.log(await exec(nargs).stdout)
+      }
+      else {
+         console.log('gridlinks package utility')
+         console.log('usage: package [patch | minor | -v | -f packagename]]')
+      }
+   }
+
+})()
